Register.tsx と BaseFrame.tsx を編集しました。
コンテキストを BaseFrame で与えるようにしました。
