+++ src/components/BaseFrame.tsx
@@ -5,24 +5,36 @@ import GlobalSpinner from "components/spinner/GlobalSpinner";
 import GlobalModal from "./modal/normal/GlobalModal";
 import GlobalModalWithInput from "./modal/with-input/GlobalModalWithInput";
 import GlobalSnackbar from "./snackbar/GlobalSnackbar";
+import GlobalSpinnerContextProvider from "contexts/spinner/GlobalSpinnerContext";
+import GlobalModalContextProvider from "contexts/modal/normal/GlobalModalContext";
+import GlobalModalWithInputContextProvider from "contexts/modal/with-input/GlobalModalWithInputContext";
+import GlobalSnackbarContextProvider from "contexts/snackbar/GlobalSnackbarContext";
 
 const BaseFrame = (props: { children: React.ReactElement }): JSX.Element => {
   const { children } = props;
   return (
-    <div>
-      <GlobalSpinner />
-      <GlobalSnackbar />
-      <Header />
-      <main
-        className={`min-h-[80vh] pb-2 bg-repeat`}
-        style={{ backgroundImage: `url(/block.png)` }}
-      >
-        {children}
-      </main>
-      <Footer />
-      <GlobalModal />
-      <GlobalModalWithInput />
-    </div>
+    <GlobalSnackbarContextProvider>
+      <GlobalModalWithInputContextProvider>
+        <GlobalModalContextProvider>
+          <GlobalSpinnerContextProvider>
+            <>
+              <Header />
+              <main
+                className={`min-h-[80vh] pb-2 bg-repeat`}
+                style={{ backgroundImage: `url(/block.png)` }}
+              >
+                {children}
+              </main>
+              <Footer />
+              <GlobalSpinner />
+              <GlobalSnackbar />
+              <GlobalModal />
+              <GlobalModalWithInput />
+            </>
+          </GlobalSpinnerContextProvider>
+        </GlobalModalContextProvider>
+      </GlobalModalWithInputContextProvider>
+    </GlobalSnackbarContextProvider>
   );
 };
